Refactor QR data logic to separate module
Moved QR code options, data structures, and related logic from main.ts to a new src/data/qr.ts module. Updated all relevant imports in components and views to use the new module, improving code organization and separation of concerns.
